ArduCopter: convert RTL altitude parameters from cm to m

+++ ArduCopter/AP_Arming.cpp
@@ -262,7 +262,7 @@ bool AP_Arming_Copter::parameter_checks(bool display_failure)
                     return false;
                 }
                 // check if RTL_ALT is higher than rangefinder's max range
-                if (copter.g.rtl_altitude > copter.rangefinder.max_distance_cm_orient(ROTATION_PITCH_270)) {
+                if (copter.g.rtl_altitude() > (int32_t)(copter.rangefinder.max_distance_cm_orient(ROTATION_PITCH_270)) ) {
                     check_failed(ARMING_CHECK_PARAMETERS, display_failure, failure_template, "RTL_ALT>RNGFND_MAX_CM");
                     return false;
                 }
